% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{best_att_subset_global}
\alias{best_att_subset_global}
\title{Attribute name stems from CFS attribute subset}
\usage{
best_att_subset_global(df, label)
}
\arguments{
\item{df}{A data frame (without class).}

\item{label}{- The class labels given as factor vector.}
}
\value{
A character vector of unique attribute name stems.
}
\description{
Performs CFS on all attributes that are available in all study waves and
returns a vector of (unique) attribute name stems.
}
