% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate a distance data frame}
\usage{
calc_dist(data, fun = "heom", trans = "range")
}
\arguments{
\item{data}{Data frame.}

\item{fun}{The distance calculation method, either as a string or as custom function.}

\item{trans}{Optional transformation before distance calculation as string or custom function.}
}
\value{
nxn distance data frame
}
\description{
Switchman function that forwards to a specific distance function or
uses a custom distance method.
}
