% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{DBSCAN Clustering on Each Study Wave}
\usage{
clustering(df, label, minPts = NULL, eps = NULL, suffix = "_s", ...)
}
\arguments{
\item{df}{The data frame. TODO: Add requirements, e.g. w.r.t. suffix}

\item{label}{The class labels given as factor vector.}

\item{minPts}{\eqn{minPts}. If not set, it defaults to \eqn{log(n)},
where \eqn{n} is the number of objects, as suggested in
\insertCite{Kailing:RIS2003}{evoxploit}.}

\item{eps}{\eqn{\epsilon}.}
}
\value{
A list containing the following elements for each wave:
\itemize{
\item \code{dist}: The \code{dist} matrix used for clustering.
\item \code{subset_att_wave}: A character vector of attribute names used
for clustering.
\item \code{kdist}: A list containing at least \eqn{minPts} and
\eqn{\epsilon}.
\item \code{clustering_result}: The DBSCAN clustering output.
}
}
\description{
Performs feature selection and DBSCAN clustering on each study wave
independently.
}
\details{
The function first employs CFS \insertCite{Hall:CFS2000}{evoxploit} to
obtain a small set of relevant,
non-redundant features for DBSCAN clustering. If for a variable only one wave
is selected (e.g. only som_bmi_s0 but not som_bmi_s1 and *._s2), it  expands
the feature space to include all possible realizations.

Then, for each wave, it applies DBSCAN. If neccessary, it uses a heuristic
to find appropriate values for DBSCAN's parameters.

Distance calculation is based on \code{\link{heom}}. If no values for
\eqn{minPts} and \eqn{\epsilon} are provided, the parameter are estimated
using a heuristic.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{heom}}, \code{\link{best_att_subset_global}},
\code{\link{kdist_info}}
}
