% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{create_CBMS15_attributes}
\alias{create_CBMS15_attributes}
\title{Generate CBMS'15+ Attributes}
\usage{
create_CBMS15_attributes(df = NULL, label = NULL, li_clustering = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{label}{The class labels given as factor vector.}

\item{li_clustering}{The output of \code{\link{clustering}}.}
}
\value{
A data frame with the following columns:
\itemize{
\item dist_s_\strong{x}_s\strong{y}: The distance of an instance to itself.
\item dev_from_pop_\strong{a}_s\strong{x}_\strong{y}: For numeric
attributes: The deviation from the population slope for \strong{a} between
\strong{x} and \strong{y}. For factors: whether the instance value AND at
least 50\% of the population have changed.
\item cluster_idx_s\strong{x}: The instance's cluster ID.
\item lof_s\strong{x}: The instance's \emph{Local Outlier Factor}
\insertCite{Breuning:LOF2000}{evoxploit}.
\item cluster_rep_s\strong{x}: Whether the instance is the cluster's
\emph{representative}. The cluster representative is the instance i, where
the sum of path distances between i and all cluster peers is minimal.
\item dist_to_rep_s\strong{x}: The instance's distance to its cluster
representative. \code{NA} if noise instance.
The distance matrix is inherited from \code{li_clustering}.
\item path_length_to_rep_s\strong{x}: The path length of direct-density
connected instances between an intance and its cluster representative.
\item dist_to_centroid_s\strong{x}: The (HEOM) distance between an instance
to its centroid.
\item dist_to_medoid_s\strong{x}: The (HEOM) distance between an instance to
its medoid.
\item frac_class_\strong{c}_in_neighborhood_s\strong{x}: fraction of
instances of class \strong{c} in the instance's \eqn{\epsilon}-neighborhood
\item silhouette_s\strong{x}: the instance's silhouette score. (noise
instances are treated as one separate cluster.)
\item real_diff_att_\strong{a}_s\strong{x}_\strong{y}:
the real difference of \strong{a} between \strong{y} and \strong{x} for an
instance. For numeric attributes only.
\item abs_diff_att_\strong{a}_s\strong{x}_\strong{y}:
the absolute difference of \strong{a} between \strong{y} and \strong{x} for
an instance. For numeric attributes only.
\item rel_diff_att_\strong{a}_s\strong{x}_\strong{y}:
the relative difference of \strong{a} between \strong{y} and \strong{x} for
an instance. For numeric attributes only.
\item has_changed_att_\strong{a}_s\strong{x}_\strong{y}:
whether the instance has changed in \strong{a} between \strong{x} and
\strong{y}. For factors only.
\item stays_outlier_s\strong{x}_\strong{y}:
whether the instance was noise in \strong{x} and remains noise in \strong{y}.
\item becomes_outlier_s\strong{x}_\strong{y}:
whether the instance was not noise in \strong{x} and becomes noise in
\strong{y}.
\item was_outlier_s\strong{x}_\strong{y}:
whether the instance was noise in \strong{x} and is not noise in \strong{y}.
\item never_outlier_s\strong{x}_\strong{y}:
whether the instance neither was noise in \strong{x} nor in \strong{y}.
\item frac_same_peers_s_\strong{x}_\strong{y}:
the fraction of cluster peers in both \strong{x} and \strong{y}.
\item same_minPts_nn_s_\strong{x}_\strong{y}:
the fraction of \eqn{minPts}-nearest neighbors in both \strong{x} and
\strong{y}.
\item real_diff_silhouette_s_\strong{x}_\strong{y}: Silhouette change.
\item real_diff_lof_s_\strong{x}_\strong{y}: LOF change.
\item diff_s_\strong{x}_\strong{y}: The instance's change w.r.t. the CFS
subset of the clustering.
}

Legend:
\itemize{
\item \strong{x} - a study wave index
\item \strong{y} - a study wave index, \strong{y} > \strong{x}
\item \strong{a} - an attribute
\item \strong{c} - a class label
}
}
\description{
Generates CBMS'15+ attributes (evolution features) from a data frame and
the output of a clustering with \code{\link{clustering}}, according to
\insertCite{Niemann:CBMS15}{evoxploit}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{create_IDA14_attributes}},
\code{\link{create_simple_attributes}}
}
