% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{create_desc_attributes}
\alias{create_desc_attributes}
\title{Generate Descriptive Temporal Attributes}
\usage{
create_desc_attributes(df, suffix, verbose = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{suffix}{A string indicating the start of the wave index suffix.}

\item{verbose}{Whether or not to show some diagnostic messages. Defaults to FALSE.}
}
\value{
A data frame with the following columns for each attribute \strong{a}
in \code{df}:
\itemize{
\item \strong{a}_desc_att_mean: The mean of \strong{a} over all waves.
For numeric attributes only.
\item \strong{a}_desc_att_sd: The st. deviation of \strong{a} over all waves.
For numeric attributes only.
#' \item \strong{a}_desc_att_median: The median of \strong{a} over all waves.
For numeric attributes only.
\item \strong{a}_desc_att_mode: The mode of \strong{a} over all waves.
\item \strong{a}_desc_att_min: The minimum of \strong{a} over all waves.
For numeric attributes only.
\item \strong{a}_desc_att_max: The maximum of \strong{a} over all waves.
For numeric attributes only.
}
}
\description{
Generates descriptive temporal attributes, e.g. the mean and mode of an attribute
over all waves.
}
\seealso{
\code{\link{create_CBMS15_attributes}},
\code{\link{create_IDA14_attributes}}
}
