% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{create_IDA14_attributes}
\alias{create_IDA14_attributes}
\title{Generate IDA'14 Attributes}
\usage{
create_IDA14_attributes(df, label, train_lgc = rep(TRUE, nrow(df)),
  suffix, kdist_sample_size = 50, seed = 123, verbose = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{label}{The class labels given as factor vector.}

\item{train_lgc}{A logical vector. \code{TRUE} elements represent training
instances of \code{df}. If not set, all instances of \code{df} are
treated as training instances.}

\item{suffix}{A string indicating the start of the wave index suffix.}

\item{kdist_sample_size}{The number of randomly sampled \eqn{epsilon} values
drawn from \code{dist_to_line} used as candidates for clustering. Weighted
random sampling is used: the larger \code{dist_to_line}, the higher the
probability of being drawn.}

\item{seed}{Random seed. Used to draw a random sample from the k-distances.}

\item{verbose}{Whether or not to show some diagnostic messages. Defaults to FALSE.}
}
\value{
A data frame. For each attribute \strong{a}, the data
frame contains an attribute \strong{a}_seq.
}
\description{
Generate sequence features as described in
\insertCite{Hielscher:IDA14}{evoxploit}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{create_CBMS15_attributes}},
\code{\link{create_simple_attributes}}
}
