% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{create_simple_attributes}
\alias{create_simple_attributes}
\title{Generate Simple Temporal Attributes}
\usage{
create_simple_attributes(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the following columns for each attribute \strong{a}
in \code{df}:
\itemize{
\item \strong{a}_desc_att_mean: The mean of \strong{a} over all waves.
For numeric attributes only.
\item \strong{a}_desc_att_sd: The st. deviation of \strong{a} over all waves.
For numeric attributes only.
#' \item \strong{a}_desc_att_median: The median of \strong{a} over all waves.
For numeric attributes only.
\item \strong{a}_desc_att_mode: The mode of \strong{a} over all waves.
\item \strong{a}_desc_att_min: The minimum of \strong{a} over all waves.
For numeric attributes only.
\item \strong{a}_desc_att_max: The maximum of \strong{a} over all waves.
For numeric attributes only.
}
}
\description{
Generates simple temporal attributes, e.g. the mean and mode of an attribute
over all waves.
}
\seealso{
\code{\link{create_CBMS15_attributes}},
\code{\link{create_IDA14_attributes}}
}
