% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{evo_permutations}
\alias{evo_permutations}
\title{Find pairwise permutations of attribute names}
\usage{
evo_permutations(df_names)
}
\arguments{
\item{df_names}{A character vector of attribute names.}
}
\value{
A list of character vectors with 2 attribute names each.
}
\description{
Find all possible pairwise wave permuations of attribute names, e.g.
{(som_bmi_s0, som_bmi_s1), (som_bmi_s0, som_bmi_s2),
(som_bmi_s1, som_bmi_s2)}. Autmatically extracts the set of wave suffixes.
}
