% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evoxploit.R
\name{Evoxploit}
\alias{Evoxploit}
\title{Object that contains original data and derived evolution features, in
addition to detailed information on the clustering.}
\format{
\code{\link{R6Class}} object.
}
\description{
\code{Evoxploit} computes evolution features on longitudinal data.
}
\section{Usage}{

\preformatted{
evo <- Evoxploit$new(data, label, wave_suffix = "_s", minPts = NULL,
eps = NULL, run = TRUE, verbose = FALSE)

summary(evo)
evo$data
evo$label
evo$evo_features
evo$all_features
evo$clu
}
}

\section{Arguments}{


For Evoxploit$new():
\describe{
\item{data: }{(`data.frame`)\cr
The data set with all input features (predictors). The object (created with Predictor$new())
holding the machine learning model and the data.}
\item{target: }{(`factor` | `numeric`) \cr
The target variable (response).}
\item{wave_suffix: }{(`character(1)`)\cr
The wave suffix given as string.}
\item{minPts: }{(`integer(1)`)\cr
(optional) minPts parameter for DBSCAN clustering.}
\item{eps: }{(`double(1)`)\cr
(optional) eps parameter for DBSCAN clustering.}
\item{run: }{(`logical(1)`)\cr
Should the (evo) features extraction method be run?}
\item{verbose: }{(`logical(1)`)\cr
Whether or not to show some diagnostic messages. Defaults to FALSE. }
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Evoxploit$new()}}
\item \href{#method-run}{\code{Evoxploit$run()}}
\item \href{#method-summary}{\code{Evoxploit$summary()}}
\item \href{#method-clone}{\code{Evoxploit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evoxploit$new(
  data,
  label,
  wave_suffix = "_s",
  minPts = NULL,
  eps = NULL,
  train_lgc = rep(TRUE, nrow(data)),
  run = TRUE,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evoxploit$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evoxploit$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evoxploit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
