% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{get_unique_waves}
\alias{get_unique_waves}
\title{Extract unique wave suffixes}
\usage{
get_unique_waves(att_names)
}
\arguments{
\item{att_names}{A character vector of attribute names.}
}
\value{
An integer vector of (sorted) wave numbers.
}
\description{
Extracts unique wave suffixes (e.g. 0 for "som_bmi_s0") from a vector of
attribute names.
}
