% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{heom}
\alias{heom}
\title{An implementation of the HEOM metric.}
\usage{
heom(data)
}
\arguments{
\item{data}{The data frame to calculate the distance matrix from.}
}
\value{
A distance matrix.
}
\description{
This creates distance matrix from given data.
}
\details{
Please note that the function doesn't scale the variables.

TODO: This breaks, if the data has only one column. Also, check if the sapply-
way to check numerics and factors is correct, since if there's only one column,
it returns logical vector for all cases, not for the column.

For further reading, see \insertCite{Wilson:HEOM1997}{evoxploit}.

Adapted from https://github.com/Tommytronic/Scatter-R/blob/master/R/heom.R.
}
\references{
\insertAllCited{}
}
