% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kdist_info}
\alias{kdist_info}
\title{k-dist graph calculations}
\usage{
kdist_info(dist_obj, k = round(log(attr(dist_obj, "Size"))))
}
\arguments{
\item{dist_obj}{A distance matrix of type \code{dist}.}

\item{k}{Used for k-dist graph. If not set, it defaults to \eqn{log(n)},
where \eqn{n} is the number of objects, as suggested in
\insertCite{Kailing:RIS2003}{evoxploit}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{knn_dist_sorted}: A double vector of sorted k-dist values.
Element names represent indices of \code{dist_obj}.
\item \code{dist_to_line}: A double vector of distances between
\code{knn_dist_sorted} and the line between its first and last element.
\item \code{idx_opt}: The index of the maximum value in \code{dist_to_line}.
\item \code{eps_opt}: The \code{idx_opt}th value in \code{knn_dist_sorted}.
\item \code{minPts}: k.
}
}
\description{
Estimates an appropriate value for the DBSCAN parameter
\eqn{\epsilon} based on the k-dist graph, as suggested in
\insertCite{Ester:DBSCAN1996}{evoxploit}.
}
\references{
\insertAllCited{}
}
