% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{range_scale}
\alias{range_scale}
\title{Scales numeric columns of a data frame}
\usage{
range_scale(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame where each numeric column is scaled.
}
\description{
Scales each numeric column of a data frame. Returns the complete
data frame.
}
