% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{range_trans}
\alias{range_trans}
\title{Range-transforms numeric columns of a data frame}
\usage{
range_trans(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame where each numeric column is range-transformed.
}
\description{
Range-transforms each numeric column of a data frame. Returns the complete
data frame.
}
