% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{replace_na_nan_inf_vec}
\alias{replace_na_nan_inf_vec}
\title{Replace NA/NaN/Inf values with a value}
\usage{
replace_na_nan_inf_vec(x, v = 0)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector with replaced values.
!!! NA -> 0
}
\description{
Replace occurences of NA/NaN/Inf values with a given value.
}
