% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxil.R
\name{replace_nan_inf_with_na}
\alias{replace_nan_inf_with_na}
\title{Replace all NaN/Inf values with NA within a data frame}
\usage{
replace_nan_inf_with_na(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
The modified data frame.
}
\description{
Replaces all occurences of NaN and Inf within a data frame with NA.
}
