% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evoxploit.R
\docType{data}
\name{Evoxploit}
\alias{Evoxploit}
\title{Object that contains original data and derived evolution features, in
addition to detailed information on the clustering.}
\format{\code{\link{R6Class}} object.}
\usage{
Evoxploit
}
\description{
\code{Evoxploit} computes evolution features on longitudinal data.
}
\section{Usage}{

\preformatted{
evo <- Evoxploit$new(data, label, wave_suffix = "_s", minPts = NULL,
eps = NULL, run = TRUE, verbose = FALSE)

summary(evo)
evo$data
evo$label
evo$evo_features
evo$all_features
evo$clu
}
}

\section{Arguments}{


For Evoxploit$new():
\describe{
\item{data: }{(`data.frame`)\cr
The data set with all input features (predictors). The object (created with Predictor$new())
holding the machine learning model and the data.}
\item{target: }{(`factor`) \cr
The target variable (response). Numeric responses
are currently not supported.}
\item{wave_suffix: }{(`character(1)`)\cr
The wave suffix given as string.}
\item{minPts: }{(`integer(1)`)\cr
(optional) minPts parameter for DBSCAN clustering.}
\item{eps: }{(`double(1)`)\cr
(optional) eps parameter for DBSCAN clustering.}
\item{run: }{(`logical(1)`)\cr
Should the (evo) features extraction method be run?}
\item{verbose: }{(`logical(1)`)\cr
Whether or not to show some diagnostic messages. Defaults to FALSE. }
}
}

\keyword{datasets}
